fixed build issues
